rft: Wpf自动战斗作业序列化调整
